Use `metrics` folder
